thcrap: Move ntdll functions to dedicated header
