+++ thcrap/src/mempatch.cpp
@@ -332,4 +332,55 @@ int vtable_detour(void **vtable, const vtable_detour_t *det, size_t det_count)
 }
 /// ----------
 
+#if TH_X64
+/// x64 manipulation of virtual memory
+/// that can be accessed with absolute encodings
+/// ----------
+static constexpr size_t low_mem_zero_mask = 0x7FFFFFFF;
+static constexpr uintptr_t null_region_size = 0x10000;
+static constexpr uint32_t max_numa_node = 0x3F;
+
+#define STATUS_SUCCESS					((DWORD)0x00000000)
+#define STATUS_INVALID_PAGE_PROTECTION	((DWORD)0xC0000045)
+#define INVALID_PARAMETER_VALUE 0x57
+
+LPVOID VirtualAllocLowEx(HANDLE hProcess, LPVOID lpAddress, SIZE_T dwSize, DWORD flAllocationType, DWORD flProtect) {
+	if (lpAddress && (uintptr_t)lpAddress < null_region_size) {
+		SetLastError(INVALID_PARAMETER_VALUE);
+		return NULL;
+	}
+	flAllocationType &= ~max_numa_node; // Make sure we don't do anything weird anyway...
+	NTSTATUS status = NtAllocateVirtualMemory(hProcess, &lpAddress, low_mem_zero_mask, &dwSize, flAllocationType, flProtect);
+	if (status < STATUS_SUCCESS) {
+		return lpAddress;
+	} else {
+		SetLastError(RtlNtStatusToDosError(status));
+		return NULL;
+	}
+}
+
+LPVOID VirtualAllocLow(LPVOID lpAddress, SIZE_T dwSize, DWORD flAllocationType, DWORD flProtect) {
+	return VirtualAllocLowEx((HANDLE)0, lpAddress, dwSize, flAllocationType, flProtect);
+}
+
+BOOL VirtualFreeLowEx(HANDLE hProcess, LPVOID lpAddress, SIZE_T dwSize, DWORD dwFreeType) {
+	if ((dwFreeType & MEM_RELEASE) && !dwSize) {
+		SetLastError(INVALID_PARAMETER_VALUE);
+		return FALSE;
+	}
+	NTSTATUS status = NtFreeVirtualMemory(hProcess, &lpAddress, &dwSize, dwFreeType);
+	if (status < STATUS_SUCCESS) {
+		return TRUE;
+	} else {
+		SetLastError(RtlNtStatusToDosError(status));
+		return FALSE;
+	}
+}
+
+BOOL VirtualFreeLow(LPVOID lpAddress, SIZE_T dwSize, DWORD dwFreeType) {
+	return VirtualFreeLowEx((HANDLE)0, lpAddress, dwSize, dwFreeType);
+}
+/// ----------
+#endif
+
 /// =============================
