+++ thcrap/src/thcrap.h
@@ -47,6 +47,7 @@ extern "C" {
 #include "exception.h"
 #include "long_double.h"
 #include "util.h"
+#include "ntdll.h"
 #include "jansson_ex.h"
 #include "expression.h"
 #include "global.h"
@@ -56,7 +57,6 @@ extern "C" {
 #include "stack.h"
 #include "binhack.h"
 #include "breakpoint.h"
-#include "memalloc_low.h"
 #include "mempatch.h"
 #include "pe.h"
 #include "plugin.h"
