+++ thcrap/thcrap.vcxproj
@@ -40,7 +40,7 @@
     <ClCompile Include="src\expression.cpp" />
     <ClCompile Include="src\fonts_charset.cpp" />
     <ClCompile Include="src\jsondata.cpp" />
-	<ClCompile Condition="'$(Platform)'=='x64'" Include="src\memalloc_low.cpp" />
+	<ClCompile Include="src\ntdll.cpp" />
     <ClCompile Include="src\mempatch.cpp" />
     <ClCompile Include="src\binhack.cpp" />
     <ClCompile Include="src\bp_file.cpp" />
@@ -85,8 +85,8 @@
     <ClInclude Include="src\fonts_charset.h" />
     <ClInclude Include="src\jsondata.h" />
     <ClInclude Include="src\long_double.h" />
-	<ClInclude Include="src\memalloc_low.h" />
     <ClInclude Include="src\mempatch.h" />
+	<ClInclude Include="src\ntdll.h" />
     <ClInclude Include="src\binhack.h" />
     <ClInclude Include="src\bp_file.h" />
     <ClInclude Include="src\breakpoint.h" />
